Add files via upload
This is a new device file for a variable speed pool water pump made by Madimack. The specific model here is called the Inverflow Pro P300i - there are a few other models in the range with more pump capacity. It is often paired with their pool heatpumps. They also have a range called Inverflow Ultra which may be compatible. Their Inverflow Eco range does not have wifi/Tuya control.

This pump has a rather obscure speed setting mechanism. There are three modes: a "boost" mode with a timer that sets the pump to 100% capacity, this is the default on startup, and the default timer is 180s. There is a "manual" mode that can set the pump from a range of 30-120%. There is also a flow based "AI" mode, that can set a pump flow rate in a variety of units. I have enabled all three modes as configurable variables, but not enumerated on the primary entity.

The primary entity I have configured as a fan (closest matching entity type in HA). The only operable aspect is the "on-off" functionality. I have added a speed DP as "read-only," which maps a read-only DP showing pump power as a range of 30-120%. This DP is the only one updated in Tuya regardless of the mode being used (manual, AI or boost). The "settable" control DP's for each of the modes otherwise act independently - i.e. if you set a speed in L/min, it does not update the %power writeable entity, and vice versa. If these control DP's are mapped to the primary fan entity, it can result in weird behaviour (i.e "speed" reads 100%, but pump is set via L/min to a lowest possible value).

There are also helpful diagnostic entity types: a boost timer countdown, current power reading, hourly reported energy use (mapped as "measurement" so it does not get added to the energy dashboard - a Riemann integral should be used instead), real flow rates and a flow pressure warning.

I have also added a select option for the flow units used by the unit. Although the DP is listed as writeable, trying to change the mode does not work on my unit (it resets back to L/min)

Because of the obscure control mechanism involving 3 modes, it is possible there is a more logical way of doing this. Let me know your feedback and I can modify. Other possible ideas I have include making the primary entity a switch only, making the speed reading a diagnostic entity. Or, using constraints to add a speed control that changes based on the mode of the pump